cleaning up code; trying to eliminate garbled LCD display
